// frontend/src/api/generated/model/extractDeckMetadataResponseDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export interface ExtractDeckMetadataResponseDto {
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  website: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  description?: string | null;
  extracted: boolean;
}