// frontend/src/api/generated/model/createStartupDtoStage.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export type CreateStartupDtoStage = typeof CreateStartupDtoStage[keyof typeof CreateStartupDtoStage];


export const CreateStartupDtoStage = {
  pre_seed: 'pre_seed',
  seed: 'seed',
  series_a: 'series_a',
  series_b: 'series_b',
  series_c: 'series_c',
  series_d: 'series_d',
  series_e: 'series_e',
  series_f_plus: 'series_f_plus',
} as const;